+++ example/src/App.tsx
@@ -8,6 +8,7 @@ import ExamplesList from './screens/ExamplesList';
 import BasicAds from './screens/BasicAds';
 import BasicAnalytics from './screens/BasicAnalytics';
 import BasicPlayback from './screens/BasicPlayback';
+import BasicTvPlayback from './screens/BasicTvPlayback';
 import BasicDrmPlayback from './screens/BasicDrmPlayback';
 import SubtitlePlayback from './screens/SubtitlePlayback';
 import ProgrammaticTrackSelection from './screens/ProgrammaticTrackSelection';
@@ -32,6 +33,7 @@ export type RootStackParamsList = {
   BasicAds: undefined;
   BasicAnalytics: undefined;
   BasicPlayback: undefined;
+  BasicTvPlayback: undefined;
   BasicDrmPlayback: undefined;
   BasicPictureInPicture: {
     navigation: NativeStackNavigationProp<RootStackParamsList>;
@@ -65,6 +67,7 @@ export type RootStackParamsList = {
 const RootStack = createNativeStackNavigator<RootStackParamsList>();
 
 const isTVOS = Platform.OS === 'ios' && Platform.isTV;
+const isAndroidTV = Platform.OS === 'android' && Platform.isTV;
 
 export default function App() {
   useEffect(() => {
@@ -117,6 +120,13 @@ export default function App() {
       },
     ],
   };
+
+  if (isAndroidTV) {
+    stackParams.data.unshift({
+      title: 'Basic TV playback',
+      routeName: 'BasicTvPlayback',
+    });
+  }
 
   if (!isTVOS) {
     stackParams.data.push({
@@ -192,6 +202,13 @@ export default function App() {
           component={BasicPlayback}
           options={{ title: 'Basic playback' }}
         />
+        {isAndroidTV && (
+          <RootStack.Screen
+            name="BasicTvPlayback"
+            component={BasicTvPlayback}
+            options={{ title: 'Basic TV playback' }}
+          />
+        )}
         <RootStack.Screen
           name="BasicDrmPlayback"
           component={BasicDrmPlayback}